separate demos into its own activity
