+++ Example/PrebidDemoJava/src/main/java/org/prebid/mobile/app/XandrBannerMoPubDemoActivity.java
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.mopub.mobileads.MoPubView;
 
+import org.prebid.mobile.AdUnit;
 import org.prebid.mobile.BannerAdUnit;
 import org.prebid.mobile.BannerBaseAdUnit;
 import org.prebid.mobile.Host;
@@ -18,8 +19,8 @@ import org.prebid.mobile.Signals;
 
 import java.util.Arrays;
 
-public class RubiconBannerMoPubDemoActivity extends AppCompatActivity {
-    BannerAdUnit adUnit;
+public class XandrBannerMoPubDemoActivity extends AppCompatActivity {
+    AdUnit adUnit;
 
     @Override
     protected void onDestroy() {
@@ -34,34 +35,31 @@ public class RubiconBannerMoPubDemoActivity extends AppCompatActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_demo);
-        PrebidMobile.setPrebidServerHost(Host.RUBICON);
-        PrebidMobile.setPrebidServerAccountId(Constants.PBS_ACCOUNT_ID_RUBICON);
-        PrebidMobile.setStoredAuctionResponse(Constants.PBS_STORED_RESPONSE_300x250_RUBICON);
-        String adSizeName = getIntent().getStringExtra(Constants.AD_SIZE_NAME);
-        int width = 0;
-        int height = 0;
+        PrebidMobile.setPrebidServerHost(Host.APPNEXUS);
+        PrebidMobile.setPrebidServerAccountId(Constants.PBS_ACCOUNT_ID_APPNEXUS);
+        BannerAdUnit adUnit = new BannerAdUnit(Constants.PBS_CONFIG_ID_300x250_APPNEXUS, 300, 250);
 
-        String[] wAndH = adSizeName.split("x");
-        width = Integer.valueOf(wAndH[0]);
-        height = Integer.valueOf(wAndH[1]);
-
-        adUnit = new BannerAdUnit(Constants.PBS_CONFIG_ID_300x250_RUBICON, width, height);
         BannerBaseAdUnit.Parameters parameters = new BannerBaseAdUnit.Parameters();
         parameters.setApi(Arrays.asList(Signals.Api.MRAID_2));
+//        parameters.setApi(Arrays.asList(new Signals.Api(5)));
+
         adUnit.setParameters(parameters);
+
+        this.adUnit = adUnit;
+//        Util.enableAdditionalFunctionality(this.adUnit);
         FrameLayout adFrame = findViewById(R.id.adFrame);
         adFrame.removeAllViews();
         final MoPubView adView = new MoPubView(this);
-        adView.setAdUnitId(Constants.MOPUB_BANNER_ADUNIT_ID_300x250_RUBICON);
-        adView.setMinimumWidth(width);
-        adView.setMinimumHeight(height);
+        adView.setAdUnitId(Constants.MOPUB_BANNER_ADUNIT_ID_300x250_APPNEXUS);
+
+        adView.setMinimumWidth(300);
+        adView.setMinimumHeight(250);
         adFrame.addView(adView);
         adUnit.setAutoRefreshPeriodMillis(getIntent().getIntExtra(Constants.AUTO_REFRESH_NAME, 0));
         adUnit.fetchDemand(adView, new OnCompleteListener() {
             @Override
             public void onComplete(ResultCode resultCode) {
                 adView.loadAd();
-
             }
         });
     }
