+++ Example/PrebidDemoJava/src/main/java/org/prebid/mobile/app/XandrNativeInBannerGamDemoActivity.java
@@ -1,0 +1,114 @@
+package org.prebid.mobile.app;
+
+import android.os.Bundle;
+import android.widget.FrameLayout;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.gms.ads.AdListener;
+import com.google.android.gms.ads.AdSize;
+import com.google.android.gms.ads.doubleclick.PublisherAdRequest;
+import com.google.android.gms.ads.doubleclick.PublisherAdView;
+
+import org.prebid.mobile.AdUnit;
+import org.prebid.mobile.Host;
+import org.prebid.mobile.LogUtil;
+import org.prebid.mobile.NativeAdUnit;
+import org.prebid.mobile.NativeDataAsset;
+import org.prebid.mobile.NativeEventTracker;
+import org.prebid.mobile.NativeImageAsset;
+import org.prebid.mobile.NativeTitleAsset;
+import org.prebid.mobile.OnCompleteListener;
+import org.prebid.mobile.PrebidMobile;
+import org.prebid.mobile.ResultCode;
+
+import java.util.ArrayList;
+
+public class XandrNativeInBannerGamDemoActivity extends AppCompatActivity {
+    AdUnit adUnit;
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        if (adUnit != null) {
+            adUnit.stopAutoRefresh();
+            adUnit = null;
+        }
+    }
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_demo);
+        PrebidMobile.setPrebidServerHost(Host.APPNEXUS);
+        PrebidMobile.setPrebidServerAccountId(Constants.PBS_ACCOUNT_ID_APPNEXUS);
+        adUnit = new NativeAdUnit(Constants.PBS_CONFIG_ID_NATIVE_APPNEXUS);
+        FrameLayout adFrame = (FrameLayout) findViewById(R.id.adFrame);
+        adFrame.removeAllViews();
+        final PublisherAdView nativeAdView = new PublisherAdView(this);
+        nativeAdView.setAdListener(new AdListener() {
+            @Override
+            public void onAdLoaded() {
+                super.onAdLoaded();
+                LogUtil.d("ad loaded");
+            }
+        });
+        nativeAdView.setAdUnitId(Constants.DFP_IN_BANNER_NATIVE_ADUNIT_ID_APPNEXUS);
+        nativeAdView.setAdSizes(AdSize.FLUID);
+        adFrame.addView(nativeAdView);
+        final PublisherAdRequest.Builder builder = new PublisherAdRequest.Builder();
+        NativeAdUnit nativeAdUnit = (NativeAdUnit) adUnit;
+        nativeAdUnit.setContextType(NativeAdUnit.CONTEXT_TYPE.SOCIAL_CENTRIC);
+        nativeAdUnit.setPlacementType(NativeAdUnit.PLACEMENTTYPE.CONTENT_FEED);
+        nativeAdUnit.setContextSubType(NativeAdUnit.CONTEXTSUBTYPE.GENERAL_SOCIAL);
+        ArrayList<NativeEventTracker.EVENT_TRACKING_METHOD> methods = new ArrayList<>();
+        methods.add(NativeEventTracker.EVENT_TRACKING_METHOD.IMAGE);
+
+        try {
+            NativeEventTracker tracker = new NativeEventTracker(NativeEventTracker.EVENT_TYPE.IMPRESSION, methods);
+            nativeAdUnit.addEventTracker(tracker);
+        } catch (Exception e) {
+            e.printStackTrace();
+
+        }
+        NativeTitleAsset title = new NativeTitleAsset();
+        title.setLength(90);
+        title.setRequired(true);
+        nativeAdUnit.addAsset(title);
+        NativeImageAsset icon = new NativeImageAsset();
+        icon.setImageType(NativeImageAsset.IMAGE_TYPE.ICON);
+        icon.setWMin(20);
+        icon.setHMin(20);
+        icon.setRequired(true);
+        nativeAdUnit.addAsset(icon);
+        NativeImageAsset image = new NativeImageAsset();
+        image.setImageType(NativeImageAsset.IMAGE_TYPE.MAIN);
+        image.setHMin(200);
+        image.setWMin(200);
+        image.setRequired(true);
+        nativeAdUnit.addAsset(image);
+        NativeDataAsset data = new NativeDataAsset();
+        data.setLen(90);
+        data.setDataType(NativeDataAsset.DATA_TYPE.SPONSORED);
+        data.setRequired(true);
+        nativeAdUnit.addAsset(data);
+        NativeDataAsset body = new NativeDataAsset();
+        body.setRequired(true);
+        body.setDataType(NativeDataAsset.DATA_TYPE.DESC);
+        nativeAdUnit.addAsset(body);
+        NativeDataAsset cta = new NativeDataAsset();
+        cta.setRequired(true);
+        cta.setDataType(NativeDataAsset.DATA_TYPE.CTATEXT);
+        nativeAdUnit.addAsset(cta);
+        int millis = getIntent().getIntExtra(Constants.AUTO_REFRESH_NAME, 0);
+        nativeAdUnit.setAutoRefreshPeriodMillis(millis);
+        nativeAdUnit.fetchDemand(builder, new OnCompleteListener() {
+            @Override
+            public void onComplete(ResultCode resultCode) {
+                PublisherAdRequest request = builder.build();
+                nativeAdView.loadAd(request);
+            }
+        });
+    }
+}
